Reference this forked repo when creating a heroku app
I need to remove the reference to the original repo, to avoid heroku using the original composer.json instead of the one in this repo. Im' doing this because app creation at heroku fails because a komodini/language package is down, so when composer tries to download it, dosen't find it and the heroku build fails.